pyproject.toml
[tool.poetry]
name = "mongo_bench"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
mongoengine = "^0.27.0"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
psycopg2 = "^2.9.5"
black = "^23.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
